+++ src/php/Compiler/Domain/Emitter/OutputEmitterInterface.php
@@ -10,7 +10,6 @@ use Phel\Compiler\Domain\Emitter\OutputEmitter\OutputEmitterOptions;
 use Phel\Compiler\Domain\Emitter\OutputEmitter\SourceMap\SourceMapState;
 use Phel\Lang\SourceLocation;
 use Phel\Lang\Symbol;
-use Phel\Lang\TypeInterface;
 
 interface OutputEmitterInterface
 {
@@ -49,7 +48,7 @@ interface OutputEmitterInterface
 
     public function emitFnWrapSuffix(?SourceLocation $sl = null): void;
 
-    public function emitLiteral(array|bool|float|int|TypeInterface|string|null $value): void;
+    public function emitLiteral(mixed $value): void;
 
     public function increaseIndentLevel(): void;